requirements.txt
black==23.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
google-api-core==2.24.2
google-api-python-client==2.108.0
google-auth==2.39.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-re2==1.0
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
importlib-metadata==6.7.0
mypy-extensions==1.0.0
numpy==1.21.6
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
proto-plus==1.26.1
protobuf==4.24.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.4
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
shortuuid==1.0.13
slack-sdk==3.26.1
tomli==2.0.1
typed-ast==1.5.5
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.7
zipp==3.15.0
